WIP: add contract-test
