+++ src/test/java/contracts/serverest/carts/CartsContract.java
@@ -1,0 +1,4 @@
+package contracts.serverest.carts;
+
+public class CartsContract {
+}
