+++ src/test/resources/schemas/users/users-schema.json
@@ -1,0 +1,42 @@
+{
+  "type": "object",
+  "properties": {
+    "quantidade": {
+      "type": "integer"
+    },
+    "usuarios": {
+      "type": "array",
+      "items": {
+        "type": "object",
+        "properties": {
+          "nome": {
+            "type": "string"
+          },
+          "email": {
+            "type": "string"
+          },
+          "password": {
+            "type": "string"
+          },
+          "administrador": {
+            "type": "boolean"
+          },
+          "_id": {
+            "type": "string"
+          }
+        },
+        "required": [
+          "nome",
+          "email",
+          "password",
+          "administrador",
+          "_id"
+        ]
+      }
+    }
+  },
+  "required": [
+    "quantidade",
+    "usuarios"
+  ]
+}
