+++ src/test/resources/schemas/users/users.json
@@ -1,0 +1,12 @@
+{
+  "quantidade": 4,
+  "usuarios": [
+    {
+      "nome": "Fulano da Silva",
+      "email": "fulano@qa.com",
+      "password": "teste",
+      "administrador": "true",
+      "_id": "0uxuPY0cbmQhpEz1"
+    }
+  ]
+}
